+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/Repositories/AcademyRepository.cs
@@ -1,10 +1,14 @@
+using System.Globalization;
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Contexts;
+using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Extensions;
 using DfE.FindInformationAcademiesTrusts.Data.Repositories.Academy;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Repositories;
 
-public class AcademyRepository(IAcademiesDbContext academiesDbContext) : IAcademyRepository
+public class AcademyRepository(IAcademiesDbContext academiesDbContext, ILogger<AcademyRepository> logger)
+    : IAcademyRepository
 {
     public async Task<AcademyDetails[]> GetAcademiesInTrustDetailsAsync(string uid)
     {
@@ -21,6 +25,68 @@ public class AcademyRepository(IAcademiesDbContext academiesDbContext) : IAcadem
             .ToArrayAsync();
     }
 
+    public async Task<AcademyOfsted[]> GetAcademiesInTrustOfstedAsync(string uid)
+    {
+        var giasGroupLinkData = await academiesDbContext.GiasGroupLinks
+            .Where(gl => gl.GroupUid == uid)
+            .Select(gl => new
+            {
+                Urn = gl.Urn!,
+                gl.EstablishmentName,
+                gl.JoinedDate
+            })
+            .ToListAsync();
+
+        var ofstedRatings = await GetOfstedRatings(giasGroupLinkData.Select(gl => int.Parse(gl.Urn)).ToArray());
+
+        var academyOfsteds = giasGroupLinkData.Select(gl =>
+                new AcademyOfsted(gl.Urn,
+                    gl.EstablishmentName,
+                    DateTime.ParseExact(gl.JoinedDate!, "dd/MM/yyyy", CultureInfo.InvariantCulture),
+                    ofstedRatings[gl.Urn].Previous,
+                    ofstedRatings[gl.Urn].Current
+                ))
+            .ToArray();
+
+        return academyOfsteds;
+    }
+
+    private async Task<Dictionary<string, OfstedRatings>> GetOfstedRatings(int[] urns)
+    {
+        // Ofsted data is held in MisEstablishments for most academies
+        var ofstedRatings =
+            await academiesDbContext.MisEstablishments
+                .Where(me => urns.Contains(me.Urn!.Value))
+                .Select(me => new OfstedRatings(me.Urn!.Value, me.OverallEffectiveness, me.InspectionEndDate,
+                    me.PreviousFullInspectionOverallEffectiveness, me.PreviousInspectionEndDate))
+                .ToListAsync();
+
+        // Look in MisFurtherEducationEstablishments for academies not found in MisEstablishments
+        // Note: if an entry is in MisEstablishments then it will not be in MisFurtherEducationEstablishments, even if it has no ofsted data
+        var urnsNotInMisEstablishments = urns.Except(ofstedRatings.Select(a => a.Urn)).ToArray();
+        if (urnsNotInMisEstablishments.Length > 0)
+        {
+            ofstedRatings.AddRange(
+                await academiesDbContext.MisFurtherEducationEstablishments
+                    .Where(mfe => urnsNotInMisEstablishments.Contains(mfe.ProviderUrn))
+                    .Select(mfe =>
+                        new OfstedRatings(mfe.ProviderUrn, mfe.OverallEffectiveness, mfe.LastDayOfInspection,
+                            mfe.PreviousOverallEffectiveness, mfe.PreviousLastDayOfInspection))
+                    .ToArrayAsync()
+            );
+        }
+
+        foreach (var urn in urns.Except(ofstedRatings.Select(a => a.Urn)))
+        {
+            logger.LogError(
+                "URN {Urn} was not found in Mis.Establishments or Mis.FurtherEducationEstablishments. This indicates a data integrity issue with the Ofsted data in Academies Db.",
+                urn);
+            ofstedRatings.Add(new OfstedRatings(urn, OfstedRating.None, OfstedRating.None));
+        }
+
+        return ofstedRatings.ToDictionary(o => o.Urn.ToString(), o => o);
+    }
+
     public async Task<int> GetNumberOfAcademiesInTrustAsync(string uid)
     {
         return await academiesDbContext.GiasGroupLinks.CountAsync(gl => gl.GroupUid == uid && gl.Urn != null);
@@ -33,5 +99,37 @@ public class AcademyRepository(IAcademiesDbContext academiesDbContext) : IAcadem
                          && gl.GroupType == "Single-academy trust")
             .Select(gl => gl.Urn)
             .FirstOrDefaultAsync();
+    }
+
+    private sealed record OfstedRatings(int Urn, OfstedRating Current, OfstedRating Previous)
+    {
+        public OfstedRatings(int urn, int? currentOverallEffectiveness, string? currentInspectionEndDate,
+            string? previousOverallEffectiveness, string? previousInspectionEndDate) :
+            this(urn,
+                currentOverallEffectiveness,
+                currentInspectionEndDate,
+                previousOverallEffectiveness is null ? null : int.Parse(previousOverallEffectiveness),
+                previousInspectionEndDate)
+        {
+        }
+
+        public OfstedRatings(int urn, int? currentOverallEffectiveness, string? currentInspectionEndDate,
+            int? previousOverallEffectiveness, string? previousInspectionEndDate)
+            : this(urn,
+                currentOverallEffectiveness == null
+                    ? OfstedRating.None
+                    : new OfstedRating(
+                        (OfstedRatingScore)currentOverallEffectiveness.Value,
+                        currentInspectionEndDate.ParseAsNullableDate()
+                    ),
+                previousOverallEffectiveness == null
+                    ? OfstedRating.None
+                    : new OfstedRating(
+                        (OfstedRatingScore)previousOverallEffectiveness,
+                        previousInspectionEndDate.ParseAsNullableDate()
+                    )
+            )
+        {
+        }
     }
 }