+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Academy/AcademyOfsted.cs
@@ -1,0 +1,9 @@
+namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Academy;
+
+public record AcademyOfsted(
+    string Urn,
+    string? EstablishmentName,
+    DateTime DateAcademyJoinedTrust,
+    OfstedRating PreviousOfstedRating,
+    OfstedRating CurrentOfstedRating
+);