+++ DfE.FindInformationAcademiesTrusts/Services/Academy/AcademyOfstedServiceModel.cs
@@ -1,0 +1,11 @@
+using DfE.FindInformationAcademiesTrusts.Data;
+
+namespace DfE.FindInformationAcademiesTrusts.Services.Academy;
+
+public record AcademyOfstedServiceModel(
+    string Urn,
+    string? EstablishmentName,
+    DateTime DateAcademyJoinedTrust,
+    OfstedRating PreviousOfstedRating,
+    OfstedRating CurrentOfstedRating
+);