+++ DfE.FindInformationAcademiesTrusts/Services/Academy/AcademyService.cs
@@ -5,6 +5,7 @@ namespace DfE.FindInformationAcademiesTrusts.Services.Academy;
 public interface IAcademyService
 {
     Task<AcademyDetailsServiceModel[]> GetAcademiesInTrustDetailsAsync(string uid);
+    Task<AcademyOfstedServiceModel[]> GetAcademiesInTrustOfstedAsync(string uid);
 }
 
 public class AcademyService(IAcademyRepository academyRepository) : IAcademyService
@@ -16,5 +17,14 @@ public class AcademyService(IAcademyRepository academyRepository) : IAcademyServ
         return academies.Select(a =>
             new AcademyDetailsServiceModel(a.Urn, a.EstablishmentName, a.LocalAuthority, a.TypeOfEstablishment,
                 a.UrbanRural)).ToArray();
+    }
+
+    public async Task<AcademyOfstedServiceModel[]> GetAcademiesInTrustOfstedAsync(string uid)
+    {
+        var academies = await academyRepository.GetAcademiesInTrustOfstedAsync(uid);
+
+        return academies.Select(a =>
+            new AcademyOfstedServiceModel(a.Urn, a.EstablishmentName, a.DateAcademyJoinedTrust, a.PreviousOfstedRating,
+                a.CurrentOfstedRating)).ToArray();
     }
 }